+++ tests/HttpClientTest.php
@@ -9,6 +9,7 @@ use DataTypes\FilePath;
 use DataTypes\Url;
 use MichaelHall\HttpClient\HttpClient;
 use MichaelHall\HttpClient\HttpClientRequest;
+use MichaelHall\HttpClient\RequestHandlers\CurlRequestHandler;
 use MichaelHall\HttpClient\Tests\Helpers\Fakes\FakeCurl;
 use PHPUnit\Framework\TestCase;
 
@@ -220,6 +221,66 @@ class HttpClientTest extends TestCase
         self::assertSame('FooBar', FakeCurl::getOption(CURLOPT_SSLCERTPASSWD));
         self::assertSame('PEM', FakeCurl::getOption(CURLOPT_SSLCERTTYPE));
         self::assertSame(FilePath::parse(__DIR__ . '/TestFiles/key.pem')->__toString(), FakeCurl::getOption(CURLOPT_SSLKEY));
+    }
+
+    /**
+     * Test fetching a page with overridden curl options.
+     */
+    public function testWithOverriddenCurlOptions()
+    {
+        $curlRequestHandler = new CurlRequestHandler();
+        $curlRequestHandler->setOption(CURLOPT_TIMEOUT, 123);
+
+        $client = new HttpClient($curlRequestHandler);
+        $request = new HttpClientRequest(Url::parse('https://example.com/'));
+        $request->addHeader('X-Test-Header: Foo Bar');
+        $response = $client->send($request);
+
+        self::assertSame(200, $response->getHttpCode());
+        self::assertTrue($response->isSuccessful());
+        self::assertSame([], $response->getHeaders());
+        self::assertSame('Hello World!', $response->getContent());
+
+        self::assertSame('https://example.com/', FakeCurl::getOption(CURLOPT_URL));
+        self::assertSame('GET', FakeCurl::getOption(CURLOPT_CUSTOMREQUEST));
+        self::assertSame(['X-Test-Header: Foo Bar'], FakeCurl::getOption(CURLOPT_HTTPHEADER));
+        self::assertNull(FakeCurl::getOption(CURLOPT_POSTFIELDS));
+        self::assertSame(123, FakeCurl::getOption(CURLOPT_TIMEOUT));
+    }
+
+    /**
+     * Test fetching a page with a curl option that is unsafe to override.
+     */
+    public function testWithUnsafeOverriddenCurlOption()
+    {
+        self::expectWarning();
+        self::expectWarningMessage('Option "CURLOPT_URL" is used internally by CurlRequestHandler. Setting it manually may lead to unexpected results.');
+
+        $curlRequestHandler = new CurlRequestHandler();
+        $curlRequestHandler->setOption(CURLOPT_URL, 'http://foobar.com/');
+    }
+
+    /**
+     * Test fetching a page with a curl option that is unsafe to override with warning suppressed.
+     */
+    public function testWithUnsafeOverriddenCurlOptionWithWarningSuppressed()
+    {
+        $curlRequestHandler = new CurlRequestHandler();
+        @$curlRequestHandler->setOption(CURLOPT_URL, 'http://foobar.com/');
+
+        $client = new HttpClient($curlRequestHandler);
+        $request = new HttpClientRequest(Url::parse('https://example.com/'));
+        $response = $client->send($request);
+
+        self::assertSame(0, $response->getHttpCode());
+        self::assertFalse($response->isSuccessful());
+        self::assertSame([], $response->getHeaders());
+        self::assertSame('Failed to connect to foobar.com: Connection refused', $response->getContent());
+
+        self::assertSame('http://foobar.com/', FakeCurl::getOption(CURLOPT_URL));
+        self::assertSame('GET', FakeCurl::getOption(CURLOPT_CUSTOMREQUEST));
+        self::assertSame([], FakeCurl::getOption(CURLOPT_HTTPHEADER));
+        self::assertNull(FakeCurl::getOption(CURLOPT_POSTFIELDS));
     }
 
     /**