+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Added setOption method to CurlRequestHandler.
 
 ## [1.3.0] - 2020-06-22
 - Added getClientCertificateType and setClientCertificateType methods to HttpClientRequest.